# Conflicts:
#	lib/common/widgets/custom_app_bar.dart
#	lib/common/widgets/custom_drawer.dart
#	lib/features/all/presentation/bloc/all_tasks_bloc.dart
#	lib/features/all/presentation/pages/all_tasks_page.dart
#	lib/features/archive/presentation/pages/archive_page.dart
#	lib/features/lists/presentation/pages/list_page.dart
#	lib/features/schedule/presentation/pages/schedule_page.dart
#	lib/features/schedule/presentation/widgets/tasks_calendar.dart
#	lib/features/settings/presentation/pages/settings_page.dart
#	lib/features/tags/presentation/pages/tag_page.dart
#	lib/features/tags/presentation/pages/tags_page.dart
#	lib/features/task_popup/presentation/views/task_popup.dart
#	localFiles/ar.json
#	localFiles/en.json

+++ lib/features/task_popup/presentation/views/task_popup.dart
@@ -18,7 +18,11 @@ import '../../../../common/dialogs/show_date_time_picker.dart';
 import '../../../../common/widgets/custom_alert_dialog.dart';
 import '../../../tasks/domain/entities/task_parameters.dart';
 
-///TODO create all day task from floating action button && control if task is all day or not
+///TODO V1.5 full page in mobile?
+
+///TODO V2 smart auto complete like Notion's / to select a list,tags,due date and start date
+///TODO V1 once start date is selected when creating task from floating button,end date is start + Globals.defaultTaskDuration
+///TODO V1.5 duration input with time
 
 // ignore: must_be_immutable
 class TaskPopupParams extends Equatable {
@@ -243,7 +247,6 @@ class TaskPopup extends StatelessWidget {
                   shape: RoundedRectangleBorder(borderRadius: borderRadius),
                   contentPadding: const EdgeInsets.all(radius),
                   actions: [
-                    ///TODO pop and push add are you sure alert dialog
                     if(task!=null)IconButton(
                         icon: Icon(
                           Icons.delete,
@@ -251,11 +254,35 @@ class TaskPopup extends StatelessWidget {
                         ),
                         onPressed: taskPopupParams.onDelete == null
                             ? null
-                            : () => taskPopupParams.onDelete!(
-                          DeleteClickupTaskParams(
-                              task: taskPopupParams.task!,
-                              clickupAccessToken: Globals
-                                  .clickupAuthAccessToken)),
+                            : () {
+                          Navigator.pop(context);
+                              showDialog(
+                            context: context,
+                            builder: (ctx) {
+                              return CustomAlertDialog(
+                                loading: false,
+                                actions: [
+                                  CustomButton(
+                                      child: Text(appLocalization.translate("delete")),
+                                      onPressed: () {
+                                        taskPopupParams.onDelete!(
+                                            DeleteClickupTaskParams(
+                                                task: taskPopupParams.task!,
+                                                clickupAccessToken: Globals
+                                                    .clickupAuthAccessToken));
+                                        Navigator.pop(ctx);
+                                      }),
+                                  CustomButton(
+                                      child: Text(appLocalization.translate("cancel")),
+                                      onPressed: () {
+                                        Navigator.pop(ctx);
+                                      }),
+                                ],
+                                content: Text(
+                                    "${appLocalization.translate("areYouSureDelete")} ${taskPopupParams.task?.name}?"),
+                              );
+                            });
+                            },
                     ),
                     CustomButton(
                         onPressed: () => Navigator.maybePop(context),
@@ -277,8 +304,7 @@ class TaskPopup extends StatelessWidget {
                       child: Column(
                         children: [
                           ///Space
-                          ///TODO B create a new Space
-                          ///TODO C default space is set in settings
+                          ///TODO V2 create a new Space
                           if(Globals.isSpaceAppWide == false)(task == null
                               ? DropdownButton<ClickupSpace>(
                                   hint:
@@ -323,7 +349,6 @@ class TaskPopup extends StatelessWidget {
                                     [],
                               ),
                               ///Priority
-                              ///TODO add priorities in case of disabled
                               if (state.isPrioritiesEnabled)
                                 DropdownButton<ClickupTaskPriority>(
                                   value: state.taskParams?.taskPriority,
@@ -399,7 +424,7 @@ class TaskPopup extends StatelessWidget {
                           ),
 
                           ///Tags
-                          ///TODO B TODO create new tags
+                          ///TODO V2 TODO create new tags
                           if(state.viewTagsButton)CustomButton(
                             customButtonEnum: CustomButtonEnum.secondary,
                               child: Text(
@@ -475,9 +500,9 @@ class TaskPopup extends StatelessWidget {
                           Wrap(
                             children: [
 
-                              ///TODO B create a new Folder
+                              ///TODO V2 create a new Folder
                               ///Folder
-                              if (state.isFoldersListAvailable && task == null)
+                              if (state.isFoldersListAvailable)
                                 DropdownButton<ClickupFolder?>(
                                   hint:
                                       Text(appLocalization.translate("folder")),
@@ -508,12 +533,9 @@ class TaskPopup extends StatelessWidget {
                               else if (task?.folder != null)
                                 Text(" ${task?.folder?.name ?? ""} "),
 
-                              ///FIXME can only select list if no list is selected
-                              ///TODO B create a new list
+                              ///TODO V2 create a new list
                               ///List
-                              if (task == null &&
-                                  state.taskParams?.clickupList == null &&
-                                  (state.taskParams?.getAvailableLists
+                              if ((state.taskParams?.getAvailableLists
                                           .isNotEmpty ==
                                       true))
                                 DropdownButton<ClickupList>(
@@ -531,9 +553,6 @@ class TaskPopup extends StatelessWidget {
                                           .toList() ??
                                       [],
                                 )
-                              else if(task == null &&
-                                  state.taskParams?.clickupList != null)
-                                Text(" ${state.taskParams?.clickupList?.name ?? ""} ")
                               else if(task!=null)
                                 Text(" ${task.list?.name ?? ""} "),
                             ],
@@ -541,7 +560,7 @@ class TaskPopup extends StatelessWidget {
 
                           Wrap(
                             children: [
-                              ///TODO is all day checkbox
+                              ///TODO V2 is all day checkbox
                               ///isAllDay
                               Checkbox(
                                   value: taskPopupParams.isAllDay,
